third_party/blink/web_tests/media/media-fragments/TC0075-expected.txt


Title: TC0075
Fragment: 't=0:01:60'
Comment: UA knows that this is an invalid media fragment, so it will play the entire media resource.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
END OF TEST
